Also added the "moved from" main thread
This essentially "moves" (theoretically/abstractly, not in any concrete
fashion that is) the main thread, and then leaves the main thread only
to do waitpid stuff. Since the singlethread implementation works just
fine, I figured I'd test this, as this could be a work around for
application that *need* to be able to be recorded by rr.

This exposes new bugs, where ptrace setoptions fail, claiming that
there's no such process.